Fixed wrong type on beatmap bpm

+++ ModAssistant/Classes/OneClickInstaller.cs
@@ -252,7 +252,7 @@ namespace ModAssistant
             public string songSubName { get; set; }
             public string songAuthorName { get; set; }
             public string levelAuthorName { get; set; }
-            public int bpm { get; set; }
+            public double bpm { get; set; }
         }
 
         public class Stats
